Add parameters to AsyncBufferedConsumer
The `AsyncBufferedConsumer` constructor now supports more arguments.
Those are the same as for the `BufferConsumer` coming from the official
Mixpanel package, which is used under the hood.

I removed the `*args` and `**kwargs` parameters so that only supported
arguments are accepted. This should avoid surprises for users.  This
breaking change is mentioned in the changelog.